reload test successfully deployed in local i2b2
